Calculate statistics for evaluating batch correction

Calculate statistics for evaluating batch correction

computeClusterEvalStats(
  spe_object,
  foiColumn,
  precomputed = NULL,
  n_dimension = c(1, 2),
  assay = 2
)

Arguments

spe_object

A Spatial Experiment object.

foiColumn

A column name indicating the factor of interest to be tested, can be biological factor or batch factor.

precomputed

a dimensional reduction results from stats::prcomp. result in reducedDims(object) to plot. Default is NULL, we will compute for you.

n_dimension

The top n dimensions to be plotted

assay

a numeric or character, specifying the assay to use (for SummarizedExperiment and its derivative classes).

Value

A dataframe object containing the clustering evaluating statistics.

Examples

library(scater)
#> Loading required package: scuttle
#> Loading required package: ggplot2
#> 
#> Attaching package: ‘scater’
#> The following objects are masked from ‘package:standR’:
#> 
#>     plotMDS, plotRLE
data("dkd_spe_subset")
spe <- scater::runPCA(dkd_spe_subset)
#> Warning: You're computing too large a percentage of total singular values, use a standard svd instead.
computeClusterEvalStats(spe, "SlideName")
#>                         types              scores
#> 1         Adjusted Rand Index   0.205918744761339
#> 2     Chi-Squared Coefficient    99.0040467838262
#> 3               Jaccard Index   0.200308166409861
#> 4 Simple Matching Coefficient                1038
#> 5             Mirkin Distance   0.785093167701863
#> 6      Silhouette Coefficient -0.0699498050976234