Import GeoMX DSP data into a saptial experiment object from file paths

Import GeoMX DSP data into a saptial experiment object from file paths

readGeoMx(
  countFile,
  sampleAnnoFile,
  featureAnnoFile = NA,
  hasNegProbe = TRUE,
  NegProbeName = "NegProbe-WTX",
  colnames.as.rownames = c("TargetName", "SegmentDisplayName", "TargetName"),
  coord.colnames = c("ROICoordinateX", "ROICoordinateY")
)

Arguments

countFile

tsv file or a dataframe object. Count matrix, with samples in columns and features/genes in rows. The first column is gene names/ids.

sampleAnnoFile

tsv file or a dataframe object. Sample annotations.

featureAnnoFile

tsv file or a dataframe object. Feature/Gene annotations.

hasNegProbe

Logical. Default is TRUE, indicating there are negative probe genes in the data.

NegProbeName

Character. Name of negative probe genes, default is NegProbe-WTX.

colnames.as.rownames

Vector of characters, length of 3. Column names used to capture gene names, sample names and gene names in countFile, sampleAnnoFile and featureAnnoFile, respectively.

coord.colnames

Vector of characters, length of 2. Column names used to capture ROI coordinates.

Value

A SpatialExperiment object.

Examples

url <- "http://nanostring-public-share.s3-website-us-west-2.amazonaws.com/"
countFile <- paste0(url, "GeoScriptHub/KidneyDataset/Kidney_Raw_TargetCountMatrix.txt")
sampleAnnoFile <- paste0(url, "GeoScriptHub/KidneyDataset/Kidney_Sample_Annotations.txt")

spe <- readGeoMx(countFile, sampleAnnoFile, hasNegProbe = FALSE)
#> Rows: 18504 Columns: 232
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: "\t"
#> chr   (1): TargetName
#> dbl (231): disease3_scan | 001 | PanCK, disease3_scan | 001 | neg, disease3_...
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> Rows: 231 Columns: 25
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: "\t"
#> chr  (8): SlideName, ScanName, SegmentLabel, SegmentDisplayName, Sample_ID, ...
#> dbl (17): ROILabel, AOISurfaceArea, AOINucleiCount, ROICoordinateX, ROICoord...
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.