Tools for analyzing NanoString's GeoMX spatial transcriptomics data

standR implements a series of functions to facilitate inspection, analysis and visualization of the NanoString's GeoMX DSP datasets. standR takes the either the csv files from the Nanostring or DGEList object as input, allowing for multiple methods to be analyzed together.

Details

standR represents the GeoMX DSP data as SpatialExperiment objects, which can easily be integrated with a wide variety of Bioconductor packages.standR generates various plots, such as QC distribution plots, dimension reduction plots and RLE plots, for quality control of genes and region of interest (ROI) samples. Multiple normalization and batch correction methods are also provided in the package as well, with the ability to compute statistics for assessing the normalization/batch correction results.

Author

Ning Liu liu.n@wehi.edu.au